Moved api call for fetching recipes to a custom hook, next to do is to create full CRUD for recipes in hook
